Item2222: Check for TML2HTML converter failures before enabling the wysiwyg editor.
If the TML2HTML converter detects a conversion failure, protect the whole topic text instead of throwing an error. That is not wysiwyg, but it does not corrupt the text, either. 
This part is needed for when the user enters something that triggers a converter failure and then clicks the WYSIWYG button.


git-svn-id: http://svn.foswiki.org/trunk@5240 0b4bb1d4-4e5a-0410-9cc4-b2b747904278