Item8911: PleaseFeelFreeToModify
This plugin doesn't work; I thought perhaps it would be an easy thing to take these guy's existing mcegooglemaps plugin for TMCE and integrate it into Foswiki. It doesn't work very well. In Foswiki it would be better to create a macro (this plugin generates lots of ugly inline <script>).

So, I am abandoning this, on the off chance it might be useful to someone.

A generic GoogleMapsPlugin? should implement a new DataForms? field type; a macro; and any TMCE integration should work with the macro instead of planting a bunch of JS into the topic (which is probably being corrupted in the HTML2TML? ). 


git-svn-id: http://svn.foswiki.org/branches/scratch@7196 0b4bb1d4-4e5a-0410-9cc4-b2b747904278