Item9235: Turns out half-arsed table clean just isn't useful
So this is a more complete approach - tries to emulate fully the
WYSIWYGPLUGIN_STICKYBITS functionality in TinyMCE so that we can
detect arbitrary HTML that contains sticky attributes, and also
clean it.

This functionality is tied to tables for now.

TODO: Rename this whole thing to TinyMCEToolsPlugin, because the code
is generic enough now to be applied to any arbitrary TinyMCE selection,
not just tables
TODO: Fix up button UI/logic so that we can apply sticky state/cleaning
to any arbitrary TinyMCE selection
TODO: Commit tablecopy code (copy selected table to clipboard as tab-delim
data)

git-svn-id: http://svn.foswiki.org/trunk@8023 0b4bb1d4-4e5a-0410-9cc4-b2b747904278