Item9338: PatternSkinUserViewTemplate is missing This view template is important Back in the old project the UserForm on top was introduced. It was a huge success! My users loves it. The user situation is that people very often need a phone number or an email address of a person while they work on a Foswiki topic. And because the wikinames are always linked to the users topic it is ONE click to get to the user topic and read a phone number or email address BUT. If the user has added stuff to their page, and many users do, you have to sometimes scroll pretty far to get to the form at the bottom
The user form is not a classification type form. It holds the most important
information on the page. 

That form was moved to the top for a very good reason. And moving it down
was not a smart move.

I am therefore moving it up again.

The same checkin that moved the form also cleaned out the all too much
garbage from the default user topic. That was a good move.
So I am only reverting form back to the top. And I will fight hard to keep it there


git-svn-id: http://svn.foswiki.org/trunk@8195 0b4bb1d4-4e5a-0410-9cc4-b2b747904278