Item9480: Fix _renderZone sometimes not flushing zones
Actually, this is a work-around for {OptimizePageLayout} = 0.

The problem was that in this case, we render both head and body zones together.

But we also call renderTML and expandMacros at the end of this process; if a
plugin has registered event handlers that add more content to either zone, then
we can no longer trust that _renderZone is going to finish with "empty" zones.

So, I simiply added a test for {OptimizePageLayout} to render the body zone,
instead of  relying on it being empty from the _renderZone(head) call.

git-svn-id: http://svn.foswiki.org/trunk@8531 0b4bb1d4-4e5a-0410-9cc4-b2b747904278