Item9553: removed excessive transclusions and repeated content. A category topic should be a "jump station" - not an admin tool of its own.
SiteMap is too heavy even for a moderate set of webs. It doesn't finish on my server within a
"normal" timeframe. So fastcgi killed it and returned a 500 internal server
error. 

SiteMap is still a problem. At least it only exists in one place and not
here and there. 

Oh, just noticed ManagingWebs is listing SiteMap as well ... rendering the Create new Web tool
rather unreachable.



git-svn-id: http://svn.foswiki.org/trunk@8754 0b4bb1d4-4e5a-0410-9cc4-b2b747904278