Item9675: statistics should default to all webs
There was no way for the statistics script to determine that the script
was called without a webname.   Foswiki.pm always sets a default web if
one is missing.    It appeared that requestedWebName was documented by
but never set by Foswiki.pm.  It is now set to the web part of the URI
string, but validated according to Topic rules.  Validating as a Topic
allows more of the "original request" to be preserved without allowing
dangerous characters to get through.

Also noted that the CLI syntax "-parm=value"  was not usable because the
order of regex meant that the whole string would match the "-parm"
regex.

git-svn-id: http://svn.foswiki.org/trunk@9169 0b4bb1d4-4e5a-0410-9cc4-b2b747904278