Item9973: Fix international chars in topic name breaking attach dialogue,
... by removing the escaping business (see code comments for justification)

Also added a new preference setting (d'oh!) for entity_encoding. The reason
is that between IRC and support web, there's at least half a dozen users
asking how to set this, and these are users who know that tinymce is capable
of such a setting and just want to know the best way to get Foswiki to play
ball.

Also, changed the default to numeric - OOB TinyMCE uses 'named'. 'raw' is
considered an advanced option (interesting ways we can break WysiwygPlugin
if you paste entities into TinyMCE that aren't supported in the charset
configured for the Foswiki site).

git-svn-id: http://svn.foswiki.org/branches/Release01x01@10483 0b4bb1d4-4e5a-0410-9cc4-b2b747904278