Item10906: Checking other than USERSWEB ??
If a username is entered into a group in a way that might cause it to be
interpreted as a Web.Topic name,  TopicUserMapper ends up with a bogus
webname when doing auth checking.  For the TopicUserMapping contrib,
user topics can't be anywhere but in the Users web - so force that name
in as the web name.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@12031 0b4bb1d4-4e5a-0410-9cc4-b2b747904278