Item11183: Sync-up with TWiki SpreadSheetPlugin Item11079: Remove leading/trailing spaces
Bug Fixes
 - Copy fix to LISTRAND
 - Copy fix for EVAL(2+08) interpreting 08 as octal. Added unit test.
 - Add unit test for Product(0,4) returning 4 -  already fixed.
 - Copy $SPLIT function,  add $nop as alias for $empty
 - Add HEXENCODE, HEXDECODE and XOR
 - Add modified BITXOR. "other" implementation really isn't a BitXOR
   - If passed a string, performs Bitwise NOT compatible with t*
   - If passed a list of integers, performs Bitwise XOR on the integers
 - Add LISTNONEMPTY function
 - Apply some refactoring,
   - Extract the recursve execution into a function
   - Change use vars to "my" definitions.
   - Prefix internal subroutines with underscore
   - Address some perlcritic messages
 - Implement DOY, WHILE() and isodate
 - Address some perlcritic warnings
 - remove regex /o option - caused some reuse issues on perl 5.12.3

In the unit test, discovered a subtle issue with LEFT (or any function
that retrieves a range of cells into a list.)  In Item9400, cells
containing a list are flattened into the list.  But the fix is in the
wrong place.  The extra split is needed on the results of _getTableRange,
not on an item that was alreay split.

Also when using a range function like LEFT, an empty cell is added
into the internal list as a single space, rather than an empty entry.
The twiki code to remove empty entries was not detecting the space as
empty.

The issue with leading/trailing space was deeper, in the routine that
extracts data from a cell.  Stripping spaces there seems to correct
multiple issues.  Removing space stripping I added elsewhere.

 - Add unit tests for COLUMN, ROW, COUNTITEMS, and COUNTSTR
LEFT and RIGHT (previously fixed) LIST bug previously fixed
SPLIT,  EVAL(2+08) interpreting 08 as octal.
HEXENCODE, HEXDECODE, XOR and BITXOR,  LISTNONEMPTY
GET/SET, ABOVE,  DEF and SETM, DOY, WHILE and isodate

git-svn-id: http://svn.foswiki.org/branches/Release01x01@12823 0b4bb1d4-4e5a-0410-9cc4-b2b747904278