Item8647: Add unit test for table foot
   * Correct generation of footer rows
   * Don't count embedded bold rows as a header or footer
   * Don't add sort links to bold cells not in header or footer

Seems to work for simple table of header, data and footer.  Changed the
code to count header and footer rows separately instead of calculating
the footer count.  Not sure if it's  correct yet for header rows that
are embedded in the middle of the table.

Rows in the middle of the table should not end up in tfoot markup.
There is still an issue here - the non-header rows still become links to
sort the table.

If the table doesn't have a header or footer, then bold cells in the
table should not be sortable.  And if the table has a footer, then it
needs to have the sort links and not bold cells elsewhere in the table.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@12867 0b4bb1d4-4e5a-0410-9cc4-b2b747904278