Item11447: Add comment_target parameter
The topic= parameter is validated early in the rest script, and if
incorrect, doesn't pass any reasonable errors back to the user.

Implement a comment_target parameter that overrides the Web.Topic for
the save so that it can be more completely validated, and return a
reasonable error to the user if there is a problem.

git-svn-id: http://svn.foswiki.org/trunk@13666 0b4bb1d4-4e5a-0410-9cc4-b2b747904278