Item11436: Global cache for HtPasswdUser:
HtPasswdUser already is caching the file, so that multiple requests in a
single session only require a single read of .htpasswd.

In previous testing, loading ~4400 users from .htpasswd required around
40ms. A topic like System/JQueryAjaxHelper hits the password manager
30 times, so having the session cache saves almost 1200ms of processing
on my test system.

Changing the cache from per-session to global is a relatively small
change, and avoids re-reading .htpasswd repeatedly which could still be
fairly significant on a very busy system.

git-svn-id: http://svn.foswiki.org/trunk@13756 0b4bb1d4-4e5a-0410-9cc4-b2b747904278