Item10238: Add unit test for Empty Groups
Verify that the admin can create an empty group.   Also delete some
extraneous user adding and assertions.  no sense creating users and
asserting they are not in a group, if they were never added to the
group.

Also noted that sessions are created as the AdminGroup rather than
AdminUser. Changed it were possible to the AdminUser, but the web
template tests fail unless the current user is the AdminGroup.

git-svn-id: http://svn.foswiki.org/trunk@14094 0b4bb1d4-4e5a-0410-9cc4-b2b747904278