Item11312: Minimal processing of HTML tables
if a <table> is encountered, it will be passed through
to the editor until the </table> is encountered.  Only processing is
blank lines are converted to <p></p> empty paragraphs.

Upon save, if the table is TML compatible, it wll be converted to TML.

This has one other side effect.   Previously HTML tables were emitted as
part of a paragraph.  With this change the beginning and trailing <p>
are no longer generated during conversion to HTML
<p>
... The table ...
</p>

This needs review!

git-svn-id: http://svn.foswiki.org/branches/Release01x01@14538 0b4bb1d4-4e5a-0410-9cc4-b2b747904278