Item11690: FormPlugin: redirect with a regular 302 instead of a 307
It is not entirely clear from SVN commit messages why the 307 was introduced in
the first place (likely during debugging Item10826). Its effect is problematic,
as noted in Item11707: A 307 causes the browser to re-send the form with a
POST, but passthrough causes the redirect target to include a
?foswiki_redirect_cache=XYZ GET parameter. Under CGI, this GET parameter is
discarded, that is, passthrough is not in effect and the target script works
with the unvalidated POST parameters (under mod_perl, the passthrough
parameters are loaded back and I suppose they overwrite POST parameters with
the same name, but who knows).

Anyway, a 307 is unnecessary. I have tested the example in Item10826 to still
work with this 302 and the fix to Item11719.

git-svn-id: http://svn.foswiki.org/trunk@14545 0b4bb1d4-4e5a-0410-9cc4-b2b747904278