Item2516: Item11872:  More fixes to blockquotes:
The issue with blockquotes "splitting" in some cases is caused when a
blockquote does not contain a paragraph. "Enter" within the blockquote
element causes it to split,  similar to newline in a list-item creating
a new item.

The solution is to generate a <p class="foswikiDeleteMe" paragraph to
enclose the blockquote contents, and remove it during save.  Also
discovered that if a blockquote is the very first element on the page,
you can't insert text before the blockquoted.   So treat blockquotes
like tables.

git-svn-id: http://svn.foswiki.org/trunk@14868 0b4bb1d4-4e5a-0410-9cc4-b2b747904278