Item12023: Add eachEvent and File::Rolling
Minor update to Logger.pm - need to require the iterator if the selected
logger does not implement it.

Added support for the File::Rolling logger, as well as cleaned up the
code a bit.  Copied the eachEventSince code from the PlainFile logger
into LogDispatch.  Works for the File logger, but needs some updates to
support File::Rolling.

git-svn-id: http://svn.foswiki.org/trunk@15285 0b4bb1d4-4e5a-0410-9cc4-b2b747904278