Item12023: Better filtering implementation
This version filters by subclassing the Log::Dispatch::File
and replacing the log() routine with a version that applies the filter.

This avoids registering and removing a separate Log::Dispatch output
class for each log attempt.

For now, the filter is a simple matching regex against the message.
This needs to be enhanced to allow individual fields in the logging hash
to be examined. By filtering the assembled message, fields like the
remote IP addres might have already been masked.

git-svn-id: http://svn.foswiki.org/trunk@15536 0b4bb1d4-4e5a-0410-9cc4-b2b747904278