Item11936: Avoid configure dying with bad paths
If bin/configure is run on a LocalSite.cfg from a different system with
completely different file system paths,  it will possibly fail to create
the .htdigest-configure file when setting the initial password.

If the directory doesn't exist, and couldn't be created, just return.  A
lot is wrong, but at least by entering configure, the there is a chance
of fixing them.

Also remove an extraneous debug statement from bin/configure.

git-svn-id: http://svn.foswiki.org/trunk@15573 0b4bb1d4-4e5a-0410-9cc4-b2b747904278