Item12196: during save, process macros in view template if $text not included
If the view template for our topic doesn't contain a template token to include
the topic text, that is probably because we don't care about the topic text at
all. Therefore, it makes sense to do the same thing as the view script does:
process/render the view template and ignore the topic text.



git-svn-id: http://svn.foswiki.org/trunk@15731 0b4bb1d4-4e5a-0410-9cc4-b2b747904278