Item12262:  Unit tests that won't fail.
I've been unable to recreate the corrupted caches.  But I've also been
unable to register any user that has a cUID with encoded characters.
Our WikiWord requirement is more restrictive than the NameFilter regex.

Maybe this could be recreated using locales?

Checked in for tinkering.  The test dumps the caches, then issues the
calls that are expected to corrupt the cache,  tests the returned cUID,
some of which ought to fail if corrupted,  and then dumps the cache
again.

git-svn-id: http://svn.foswiki.org/trunk@16088 0b4bb1d4-4e5a-0410-9cc4-b2b747904278