Item12605: Eliminate extra posts for non-autocheck
It appears that all fields resulted in a post / feedback request
regardless of the setting of FEEDBACK=AUTO in config.spec.  This seems
to be some redundant code, since the AUTO fields have an on-click
handler installed to handle the live check.

Note, this increments the version of scripts.js.    Developers should
either re-pseudo-install, or run lib/Foswiki/Configure/resources/make_gz
in that directory.  And then refresh the page, or clear the browser
cache.

There is still a performance issue on chrome browser when a post is
made.

git-svn-id: http://svn.foswiki.org/trunk@16938 0b4bb1d4-4e5a-0410-9cc4-b2b747904278